Make `parameters::{Method,Call}` directly accessible
