+++ monument/lib/src/group.rs
@@ -44,6 +44,11 @@ impl PartHeadGroup {
         }
     }
 
+    /// Create a new `PartHeadGroup` containing only one part.
+    pub fn one_part(stage: Stage) -> Self {
+        Self::new(&RowBuf::rounds(stage))
+    }
+
     /// Return the maximum [`effective_stage`](Row::effective_stage) of every [`Row`] in this
     /// `PartHeadGroup`
     pub fn effective_stage(&self) -> Stage {
